Move one-liners to index.tsx
