--- frontend/packages/console-shared/src/components/status/ErrorStatus.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { RedExclamationCircleIcon } from './Icons';
-import { StatusComponentProps } from './types';
-import GenericStatus from './GenericStatus';
-
-const ErrorStatus: React.FC<StatusComponentProps> = (props) => (
-  <GenericStatus {...props} Icon={RedExclamationCircleIcon} />
-);
-
-export default ErrorStatus;
