--- frontend/packages/console-shared/src/components/status/index.ts
@@ -1,12 +1,0 @@
-export { default as Status } from './Status';
-export { default as StatusIcon } from './StatusIcon';
-export { default as StatusIconAndText } from './StatusIconAndText';
-export { default as PopoverStatus } from './PopoverStatus';
-export { default as LinkStatus } from './LinkStatus';
-export { default as SuccessStatus } from './SuccessStatus';
-export { default as ErrorStatus } from './ErrorStatus';
-export { default as InfoStatus } from './InfoStatus';
-export { default as ProgressStatus } from './ProgressStatus';
-export { default as SecondaryStatus } from './SecondaryStatus';
-export { default as PendingStatus } from './PendingStatus';
-export * from './Icons';
