+++ frontend/packages/console-shared/src/components/status/index.tsx
@@ -1,0 +1,8 @@
+export * from './icons';
+export * from './statuses';
+export * from './Status';
+
+export { default as SecondaryStatus } from './SecondaryStatus';
+export { default as LinkStatus } from './LinkStatus';
+export { default as StatusIconAndText } from './StatusIconAndText';
+export { default as PopoverStatus } from './PopoverStatus';
